signac 2.0 changes

+++ tests/test_synced_collections/test_redis_collection.py
@@ -7,20 +7,21 @@ import uuid
 import pytest
 from synced_collection_test import SyncedDictTest, SyncedListTest
 
-from signac.synced_collections.backends.collection_redis import RedisDict, RedisList
+from signac._synced_collections.backends.collection_redis import RedisDict, RedisList
 
 try:
     import redis
 
     try:
-        # try to connect to server
+        # Try to connect to server
         redis_client = redis.Redis()
         test_key = str(uuid.uuid4())
         redis_client.set(test_key, 0)
-        assert redis_client.get(test_key) == b"0"  # redis stores data as bytes
+        if redis_client.get(test_key) == b"0":  # Redis stores data as bytes
+            raise RuntimeError("Cache access check failed.")
         redis_client.delete(test_key)
         REDIS = True
-    except (redis.exceptions.ConnectionError, AssertionError):
+    except (redis.exceptions.ConnectionError, RuntimeError):
         REDIS = False
 except ImportError:
     REDIS = False
